+++ src/pages/dao/tabs/index.js
@@ -1,3 +1,3 @@
-export { DaoParameters } from './Parameters'
-export { DaoProposals } from './Proposals'
-export { SubmitDaoProposals } from './Submit'
+export { default as DaoParameters } from './Parameters'
+export { default as DaoProposals } from './Proposals'
+export { default as SubmitDaoProposals } from './Submit'
